refactor: Rename hook to useCampaignEntityRecord
