+++ src/Campaigns/resources/admin/components/CampaignDetailsPage/useCampaignEntityRecord.tsx
